harvester-app/requirements.txt
pytesseract 
pillow
numpy
opencv-python
# pytubefix
# pytube
yt-dlp
black
isort
flake8